// src/lib/generated/LocationValue.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Сколько единиц и на какую сумму лежит на каждом складе.
 */
export type LocationValue = { locationId: number, location: string, items: number, units: number, value: number, };